rename card component
