+++ frontend/src/pages/payments/index.tsx
@@ -1,13 +1,8 @@
 import React from "react";
-import Layout from "../../components/Layout";
 import PaymentsPage from "../../containers/PaymentsPage";
 
 const Payment = () => {
-  return (
-    <Layout>
-      <PaymentsPage />
-    </Layout>
-  );
+  return <PaymentsPage />;
 };
 
 export default Payment;
